Refactor Command Event

+++ src/snorlax/event.c
@@ -68,7 +68,7 @@ extern void event_func_on(___notnull event_t * event) {
 #endif // RELEASE
 
     object_lock(event->subscription);
-    event_subscription_on(event->subscription, event->type, (uint64_t) event->node);
+    event_subscription_on(event->subscription, event->type, event->node);
     object_unlock(event->subscription);
 
     event_rem(event);
